refactor(waymo-converter): separate Waymo-derived and NVIDIA-original code
Split utils.py into three files with clear license boundaries:

- proto_utils.py (NVIDIA-only): fast protobuf wire-format parser for
  MatrixFloat/MatrixInt32 protos, replacing slow ParseFromString
- waymo_range_image_utils.py (Waymo Apache-2.0 + NVIDIA modifications):
  modified range image functions (polar, cartesian, point cloud extraction
  with timestamps/azimuths/ray origins) derived from upstream v1.5.1
- utils.py (NVIDIA-only): pose extrapolation and batched multi-return
  point cloud conversion with no Waymo-derived code

Verbatim Waymo functions (compute_inclination, get_rotation_matrix,
get_transform, _combined_static_and_dynamic_shape) are now imported
directly from the upstream waymo_open_dataset package instead of being
vendored, enabled by a new Bazel patch that remaps @wod_deps to
@ncore_pip_deps in the upstream utils/BUILD.
